pyproject.toml
[tool.poetry]
name = "todoapi"
version = "0.1.0"
description = "backend for todo application"
authors = ["Igor Tereschuk <igortereschuk92@gmail.com>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
selenium = "^4.9.1"
psycopg = {extras = ["c"], version = "^3.1.9"}
requests = "^2.31.0"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
